+++ app/utils/draft/index.ts
@@ -3,7 +3,7 @@
 
 import {Alert, type AlertButton} from 'react-native';
 
-import {removeDraft} from '@actions/local/draft';
+import {parseMarkdownImages, removeDraft, updateDraftMarkdownImageMetadata, updateDraftMessage} from '@actions/local/draft';
 import {General} from '@constants';
 import {CODE_REGEX} from '@constants/autocomplete';
 import {t} from '@i18n';
@@ -172,6 +172,26 @@ export function alertSlashCommandFailed(intl: IntlShape, error: string) {
         error,
     );
 }
+
+export const handleDraftUpdate = async ({
+    serverUrl,
+    channelId,
+    rootId,
+    value,
+}: {
+    serverUrl: string;
+    channelId: string;
+    rootId: string;
+    value: string;
+}) => {
+    await updateDraftMessage(serverUrl, channelId, rootId, value);
+    const imageMetadata: Dictionary<PostImage | undefined> = {};
+    await parseMarkdownImages(value, imageMetadata);
+
+    if (Object.keys(imageMetadata).length !== 0) {
+        updateDraftMarkdownImageMetadata({serverUrl, channelId, rootId, imageMetadata});
+    }
+};
 
 export function deleteDraftConfirmation({intl, serverUrl, channelId, rootId, swipeable}: {
     intl: IntlShape;
